Update readme to fit purpose of fork
All credit for the original program to it's developers.  Miyamoto!NX is intended to provide an open source, quickly released NSMBUDX level editor and not to steal code :)  If I'm somehow trying to push these changes to the original Miyamoto! repository please let me know and I'll just remake it from scratch instead of forking!